requirements-file.txt
flask==2.3.3
web3==6.11.0
solana==0.30.0
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.9.0
pycryptodome==3.19.0
pandas==2.1.0
matplotlib==3.8.0
pillow==10.2.0
cairosvg==2.7.1
base58==2.1.1
pynacl==1.5.0
openpyxl==3.1.2